feat: reconcile watched feeds on reconnect (#54)
This PR focuses on reconciling and syncing feeds whenever we reconnect.
Since we do not have any more advanced mechanisms to do this as of now
(no `sync` API or specific `watch` endpoints) we're keeping it
relatively simple, doing the following:

- Every time `feed.getOrCreate` is invoked without pagination
(`request.next` is `undefined`), we store the entire `request`
configuration
- The next time we receive `connection.changed` event with `online:
true`, we go through all of `this.activeFeeds` and we:
  - Run `feed.getOrCreate` for each `feed` that was previously `watch`ed
- We do it with the last saved configuration, since we don't want to be
doing it on a pagination `request`
- We intentionally skip the very first `connection.changed` as this will
always fire immediately as the connection is established; which means
we'll likely be querying twice for the initial set of feeds we acquire

The PR also applies a workaround for pagination of `activities`, which
is now working but we're hitting an odd edge case explained [here for
more
context](https://getstream.slack.com/archives/C06RK9WCR09/p1752249505895919).