feat: search implementation and bindings (#78)
🎫 Ticket: https://linear.app/stream/issue/XYZ-123

📑 Docs: https://github.com/GetStream/docs-content/pull/<id>

### 💡 Overview

This PR implements basic search functionality in the RN sample app as
well as adding utilities for it in the `@react-bindings` package, mostly
related to:

- Search controller state
- Search source state
- General searching utilities

It also addresses 2 bugs in the search implementation and introduces a
new feature which allows passing an empty query string in the search
sources (removing all `filters` or arguments related to it in the
process). This is useful in a presearch scenario, where you'd like to
have something loaded and apply search to it only after.

### 📝 Implementation notes