feat: add unhandled error events (#109)
🎫 [Ticket:
https://linear.app/stream/issue/XYZ-123](https://linear.app/stream/issue/REACT-557/emit-a-local-event-if-connection-recovery-failed)

📑 Docs:[
https://github.com/GetStream/docs-content/pull/<id>](https://github.com/GetStream/docs-content/pull/585)

### 💡 Overview

This PR adds a custom local event that emits unhandled errors in the
LLC.

For now, the only one of these is the reconnection reconciliation event
as after some thorough searching I found it to be the only place where
this might happen.

I was able to get the LLC in a "broken" state on failing various calls
to `loadToken` under the hood (if token refresh is not working properly
for example), but it could only happen if the `tokenProvider` is broken
and I deemed that something that should be handled upstream, in the
integrator's code.

We can add it if we decide to do so in the future.

### 📝 Implementation notes