feat: reaction state updates on http responses (#118)
🎫 Ticket:
https://linear.app/stream/issue/REACT-561/update-reaction-and-comment-state-without-ws-events

📑 Docs: https://github.com/GetStream/docs-content/pull/<id>

### 💡 Overview

This PR is part of a bigger refactoring around updating state from HTTP
responses. It covers activity reactions and some follow refactoring for
now.

In this set of changes, it includes:

- Implementing an `activity` ID cache on each `Feed` to make sure
updates happen only on `feed`s where they're needed (particularly
important on HTTP response state updates or receiving WS events for
activities that aren't part of any of the `activities` pages)
- Implements HTTP response state updates for adding and removing
reactions to activities
- Extends @szuperaz 's [(awesome)
implementation](https://github.com/GetStream/stream-feeds-js/pull/75) of
our `stateUpdateQueue` so that:
- Any payloads that the current'y `connected_user` has not triggered are
not added to the queue (but their updates are accepted by default)
- We differentiate between WS or HTTP payloads, so that each one cancels
each other out (this solves bugs such that if a WS event never arrives
and we try triggering the HTTP request again, since the queue will never
be flushed we won't be able to do it at all the first time around; i.e
in `for-you` feeds)
- A bunch of tests

For a more in-depth explanation of how the queue works, please refer to
the JSDocs added to `shouldUpdateState`.

### 📝 Implementation notes

In case of foryou-style feeds WS event won't arrive even though we watch
them. In these cases items will persist in state update queue. Since
this is only for foryou-style pages, and users's own event shouldn't be
more than a couple, it's ok. If we'll have more lingering items in
future, we'll need to add a cleanup mechanism that removed old items.