feat: state updates on http responses for comments (#120)
🎫 Ticket:
https://linear.app/stream/issue/REACT-561/update-reaction-and-comment-state-without-ws-events

📑 Docs: https://github.com/GetStream/docs-content/pull/<id>

### 💡 Overview

This is a continuation of [this
PR](https://github.com/GetStream/stream-feeds-js/pull/118) that
finalizes the basic functionality around HTTP response updates. There
are still some direct HTTP requests missing (that do not trigger state
updates implicitly) however these can be added later on as we should now
have a stable framework to do so.

In this PR we handle:

- Adding comment reactions
- Removing comment reactions
- Adding comments
- Removing comments
- Updating comments
- Updating parent comments where necessary
- Updating parent activities where necessary

For now, we only update `comment_count` in the parent `activity` if
`watch: false`. We can add the state update queueing mechanism to
`feeds.activity.updated` as well in a separate change which should allow
us to freely do `activity` updates regardless of the `watch` state.

The 2 `FIXME`'s have already been addressed by the BE team (thank you
@kmitrovv) and will be added to this PR once deployed.

TODO: 

- [x] Implement changes to reflect work on the 2 `FIXME`'s that have
already been merged by the backend team (pending deployment)

### 📝 Implementation notes