feat: add support for unique reactions (#130)
🎫 Ticket:
https://linear.app/stream/issue/REACT-581/add-support-for-enforcing-unique-reactions

📑 Docs: https://github.com/GetStream/docs-content/pull/<id>

### 💡 Overview

This PR adds support for unique reactions within the SDK.

Unique reactions always trigger the `comment.reaction.updated` or
`activity.reaction.updated` events so we can use them as some
fine-grained control to determine which reaction action is happening. We
mainly need this in order to modify `own_reactions` adequately on the
parent `entity`, as it won't arrive through WS events.

Since this is a preliminary PR before this is deployed on our BE, we
have to wait for [this PR](https://github.com/GetStream/chat/pull/10143)
to be deployed before merging this one.

The PR's been tested against WS events, HTTP response state updates (and
of course both).

TODO:

- [x] Tests
- [x] Docs
- [x] (optional, probably better to do it later) Clean up the reaction
actions to eliminate code duplication

### 📝 Implementation notes

---------

Co-authored-by: Zita Szupera <szuperaz@gmail.com>