fix: issue with commonjs exports (#140)
🎫 Ticket:
https://linear.app/stream/issue/RN-288/react-native-sdk-is-erroring-out

📑 Docs: https://github.com/GetStream/docs-content/pull/<id>

### 💡 Overview

This PR fixes an issue we have in `feeds-client` specifically for React
Native projects that do not yet support `exports`.

An easy way to test this is to do something like
`config.resolver.unstable_enablePackageExports = false;` in your
`metro.config.js` file and it will fail. Even though
`unstable_enablePackageExports` is available in the versions we
currently support for React Native on the SDK, it is not guaranteed that
libraries compatible with those versions specifically will also support
`exports` and be declared `exports` compatible. This is especially
available on `Expo` apps, where the default version of their `Metro`
config has this turned off (until RN `0.79.0` and onwards).

Until `exports` are widely accepted and consumed, we'll rely on these
proxy files that directly create the subpaths needed for
`import`s/`require`s.

If `exports` are supported, the defined `./react-bindings` entrypoint
will always be respected.

### 📝 Implementation notes