feat: watch activity (#149)
🎫 Ticket: https://linear.app/stream/issue/REACT-645/watch-activities

📑 Docs: https://github.com/GetStream/docs-content/pull/743/files

### 💡 Overview

For activity detail pages, integrators need a way to create a reactive
state store for a single activity. To solve this issue,
`ActivityWithStateUpdates` is introduced.

Public API is shown in docs:
https://github.com/GetStream/docs-content/pull/743/files

### 📝 Implementation notes

#### `ActivityWithStateUpdates`

What it does:
- `activity.get` method loads the activity, and optionally comments as
well
- Under the hood, it creates a `Feed` instance that takes care of state
updates
- It's up to integrators to watch the feed; if the feed is not watched,
the activity state is still updated on HTTP responses
- Since an activity can be posted to multiple feeds (directly, timelines
don't cause issue), and `ActivityWithStateUpdates` can't know which of
the containing feeds are watched, we update the activity if a WS event
is received with any of the containing feeds' `fid`
- A known limitation is that if a feed is posted to multiple feeds
(timelines are not a problem, only if an activity is directly posted to
multiple feeds), and the integrator watches more than one of them, the
duplicate state updates can cause issues. It's a tricky issue to solve,
but it's an edge-case. I'll create a follow-up ticket, but I didn't want
to blow this task any bigger.
- `activity.synchronize` will reload the activity if any of the
containing feeds are watched
- I thought a lot about the naming, at first it was `Activity`, but I
realized that in the sample app, we already use the `activity` name for
`ActivityResponse` variables, so it got confusing. Instead, I opted for
`activityWithStateUpdate = client.activityWithStateUpdates(...)`

#### `useActivityComments`

Supports reading comments via `feed` or `activityWithStateUpdates`

The existing `useComments` hook is deprecated. It's because it used the
`parent` term differently than the Stream API. It is called `parent`
either the activity or the parent comment of a comment. It's logical,
but in the API, `parent` can only mean parent comment.

#### React sample app
- Create activity details page
- Created multiple small components from the `Activity` component
- `Activity` component is used when we read an activity inside a feed
- `Activity` page is used when displaying a single activity